backend/requirements.txt
# Magnes Backend Requirements

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# AI / LLM
langgraph>=0.0.50
langchain-core>=0.1.0
langchain-openai>=0.0.5
openai>=1.6.0

# RAG
llama-index>=0.9.0
chromadb>=0.4.0
ragas>=0.1.0
rank-bm25>=0.2.2

# Browser Automation
playwright>=1.40.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.25.0
requests>=2.31.0

# Image Processing
Pillow>=10.1.0

# NLP / Text Processing
jieba>=0.42.1
beautifulsoup4>=4.12.0

# MCP
mcp>=0.1.0

# Utilities
uuid>=1.30
datasets>=2.14.0
